requirements.txt
# Database
sqlalchemy>=2.0.0

# HTTP / APIs
openai>=1.0.0
google-generativeai>=0.4.0
anthropic>=0.20.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# PDF generation
reportlab>=4.0.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Utilities
python-dateutil>=2.8.0
